Permits: Move factory functions to factories.permit
These functions belong to the factories.  This move also removes the
cyclic dependency.
