+++ parkings/tests/api/enforcement/test_permit.py
@@ -3,8 +3,9 @@ from django.core.urlresolvers import reverse
 from rest_framework.status import (
     HTTP_200_OK, HTTP_201_CREATED, HTTP_400_BAD_REQUEST, HTTP_403_FORBIDDEN)
 
+from ....factories.permit import (
+    generate_areas, generate_external_ids, generate_subjects)
 from ....models import PermitCacheItem
-from ...utils import generate_areas, generate_external_ids, generate_subjects
 
 list_url = reverse('enforcement:v1:permit-list')
